requirements.txt
torch
torchvision
gradio
matplotlib
tqdm
opencv-python
scipy
einops
trimesh
huggingface-hub[torch]>=0.22
imageio[ffmpeg]
mini-dust3r